toggle to use original discovery code with featureFlag
